fix: test case corrections
