+++ tests/contrib/sqlalchemy/repository/test_sqlalchemy_asyncpg.py
@@ -2,8 +2,7 @@
 from __future__ import annotations
 
 import sys
-from asyncio import AbstractEventLoop, get_event_loop_policy
-from typing import Any, AsyncGenerator, Iterator
+from typing import Any, AsyncGenerator
 
 import pytest
 from sqlalchemy import NullPool
@@ -27,16 +26,6 @@ pytestmark = [
     pytest.mark.skipif(sys.platform != "linux", reason="docker not available on this platform"),
     pytest.mark.usefixtures("postgres_service"),
 ]
-
-
-@pytest.fixture(scope="session")
-def event_loop() -> Iterator[AbstractEventLoop]:
-    """Need the event loop scoped to the session so that we can use it to check
-    containers are ready in session scoped containers fixture."""
-    policy = get_event_loop_policy()
-    loop = policy.new_event_loop()
-    yield loop
-    loop.close()
 
 
 @pytest.mark.sqlalchemy_asyncpg
