+++ tests/contrib/sqlalchemy/repository/test_sqlalchemy_psycopg_sync.py
@@ -2,8 +2,7 @@
 from __future__ import annotations
 
 import sys
-from asyncio import AbstractEventLoop, get_event_loop_policy
-from typing import Any, Generator, Iterator
+from typing import Any, Generator
 
 import pytest
 from sqlalchemy import Engine, NullPool, create_engine
@@ -22,16 +21,6 @@ pytestmark = [
     pytest.mark.skipif(sys.platform != "linux", reason="docker not available on this platform"),
     pytest.mark.usefixtures("postgres_service"),
 ]
-
-
-@pytest.fixture(scope="session")
-def event_loop() -> Iterator[AbstractEventLoop]:
-    """Need the event loop scoped to the session so that we can use it to check
-    containers are ready in session scoped containers fixture."""
-    policy = get_event_loop_policy()
-    loop = policy.new_event_loop()
-    yield loop
-    loop.close()
 
 
 @pytest.mark.sqlalchemy_psycopg_sync
